+++ src/Controllers/Tags.hs
@@ -31,7 +31,7 @@ import Responses
     , responseOKJSON
     , responseOk
     )
-import Types (TokenLifeTime)
+import Types.Other (TokenLifeTime)
 
 sendTagsList ::
        MonadIO m
